+++ apps/forge-web/src/components/withRoot.js
@@ -5,6 +5,7 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import { colors, sizes } from '../libs/constant';
 import { useThemeMode } from '../libs/hooks';
 
+// https://material-ui.com/customization/default-theme/
 const createTheme = mode =>
   createMuiTheme({
     palette: {
@@ -14,6 +15,7 @@ const createTheme = mode =>
         paper: mode === 'light' ? colors.white : colors.gray,
         default: mode === 'light' ? colors.background : colors.gray,
       },
+      type: mode,
     },
     typography: {
       useNextVariants: true,