+++ apps/forge-web/src/pages/explorer/components/tx_card/account_migrate.js
@@ -86,6 +86,8 @@ const Container = styled.div`
     overflow: hidden;
     white-space: no-wrap;
     text-overflow: ellipsis;
+    color: ${props => props.theme.typography.color.gray};
+    font-size: 14px;
 
     p {
       text-align: left;
@@ -106,6 +108,7 @@ const Container = styled.div`
   .meta {
     font-size: 14px;
     font-weight: 600;
+    color: ${props => props.theme.typography.color.gray};
     .meta-icon,
     .meta-separator {
       margin-right: 8px;