+++ apps/forge-web/src/pages/explorer/components/tx_card/consume_asset.js
@@ -83,9 +83,11 @@ const Container = styled.div`
     text-overflow: ellipsis;
     text-align: left;
     font-size: 14px;
+    color: ${props => props.theme.typography.color.gray};
 
     a {
       color: ${props => props.theme.typography.color.main};
+      font-size: 14px;
     }
 
     .type {